Changes searchType requirements.
searchType is no longer required in the CreateSearch input type as
sometimes we want to be able to create "searches" without knowing the
type yet. This means that we also have to allow searchType in the
UpdateSearch input type, since we will need to be able to change it at a
later date. All of this is in aid of allowing people to copy a search
and set a new mask or target library.