Fixes broken mask & search when logged in via google
DynamoDB uses the value in the owner field to determine access to the
search records. When creating the new search after copying the image to
be masked, the lambda function was using the users 'sub' to set the
owner. For AWS logins, this is fine, but when logged in with a google
account, the users sub is not the same as the username. Switching to
username seems to get the code working for both types of logins.