Fixed cases where dynamoDB queries or RegExp were failing
There are some neuron instances that have non alphanumeric characters in
them, eg: (AVM06)_R. These cause the regular expression that checks the
returned matches to fail. An escape function was added to the code to
make sure these values are no longer an issue.

Some of the dynamo DB queries were failing, because the supplied
variables were incorrect when multiple wildcards were in the string, eg:
*AVM*_L was not able to generate a valid query. This was fixed in the
published_names code.