Renamed dynamoDB searchTable.
This table needs to be renamed, so that the data migration from the
existing searches to the newly released production site will work. If we
don't rename the table, then the records wont be copied correctly. This
should only need to be done for release 2.0.0. Once all the older
records have been migrated or "enough" time has passed, the old table
can be removed.