Reverts download creator lambda back to node v12
For some reason on node v14 the lambda was choosing a uuid when
requested, creating the zip file, but then not writing it to disk until
the next call of the lambda function. The requester would get the new
uuid in the response, go to that location in the s3 bucket and find
nothing.