Modifies Cognito pool to correctly assign admin.
The admin role was not being assigned correctly since it was added. This
meant that the admin user was unable to see the private search data for
other users after the permissions on the search data bucket were
updated. This fix makes sure that the admin role is assigned if it is
found in the users token when authenticating.