Adds search migration lambda function.
This function can be called from the aws console to convert custom
searches in the dynamodb searches table for the current environment. All
searches that have reached step 4 and have a .results file will first
backed up to a .results.2.0.back file and then converted to the new 3.0
data model. If a json structure of { "revert": true } is passed to the
function, it will replace the newly created .results files with the
backup files that were generated during the conversion.