Adds uploaded format check to channel creation lambda
This lambda was assuming that a 2D image was being uploaded. If it was
anything else, the code would timeout trying to load the image. The new
check will throw the correct error about the image type and let the user
know that they need to alter their uploaded image or search type.