fix: removes user/data migration code.
This iss legacy code that was used to migrate user accounts from a pre
existing cognito pool to the one currently in use. It was set up 3 years
ago and anyone that hasn't logged in for the last three years, can
probably deal with creating a new account. Once removed, this should
allow us to remove the legacy cognito pool that was required to
determine if a migration was needed.