fix: Adds listItemsByOwner query to appsync api.
This allows us to filter and list searches based on a global secondary
index on the searches table. This index uses less search tokens and
returns the correct number of results as opposed to having the filtered
after the query has been run. This should allow for pagination as a next
step.