fix: Adds additional search parameters to search_copy
This allows us to copy the originally selected search parameters when
using the 're-select mask' button on the search uploads page. These
parameters can then be used to populate the form on the mask selection
page, so that the person running the re-selection doesn't need to
remember the choices they made previously.